Refactor React-Router Code to use createBrowserRouter
Useful Link: https://reactrouter.com/en/main/routers/create-browser-router

According to the react-router-devs, it's best practice to use createBrowserRouter to render all of our routes. This means that we now don't need links.ts or any of the types/interfaces it was using. We now use a RouterProvider with all the routes being in Router.tsx. This also meant creating a Root.tsx and setting the others like Login, Logout, and Home as its children.

Refactoring: All the provider code was moved to App.tsx to clear out index.tsx.

NotFound.tsx is now in router directory since it is rendered as an errorElement.